CI/k8s: Change kafka image to point to a more stable tag 1.1.0 instead of latest
This commit is related to the CI flake: K8sValidatedKafkaPolicyTest KafkaPolicies test failed because kafka image went into error state.
This change tries to use a more stable kafka docker image tag 1.1.0 instead of latest. This will help to avoid breaking the CI like this
if something changes in latest branch of the kafka image.

Fixes: #4431
Signed-Off-By: Manali Bhutiyani <manali@covalent.io>